laboratorio/requirements.txt
agate==1.6.0
agate-dbf==0.2.0
agate-excel==0.2.3
agate-sql==0.5.2
alabaster==0.7.8
ansible-core==2.13.13
apache-libcloud==2.8.0
apturl==0.5.2
argcomplete==1.8.1
arrow==1.3.0
astroid==3.0.1
asttokens==2.4.1
atlassian-python-api==3.41.13
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
awscli==1.29.65
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
binance-connector==3.5.1
binaryornot==0.4.4
black==23.11.0
bleach==3.1.1
blinker==1.4
boto3==1.26.155
botocore==1.31.65
Brlapi==0.7.0
cached-property==1.5.1
capstone==4.0.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
comm==0.2.0
command-not-found==0.3
contourpy==1.1.1
cookiecutter==2.5.0
coverage==7.5.3
cryptography==2.8
csvkit==1.0.2
cupshelpers==1.0
cycler==0.12.1
dbfread==2.0.7
dbus-python==1.2.16
debugpy==1.8.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docstring-to-markdown==0.13
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
flake8==6.1.0
fonttools==4.46.0
future==0.18.2
h11==0.14.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.27.1
ipython==8.12.3
ipython_genutils==0.2.0
isodate==0.6.0
isort==4.3.4
jdcal==1.0
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==3.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
leather==0.3.3
leechcorepyc==2.14.3
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.8
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.24.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
outcome==1.3.0.post0
packaging==20.3
pandas==2.0.3
pandocfilters==1.4.2
paramiko==2.6.0
parsedatetime==2.4
parso==0.8.3
pathspec==0.11.2
pbr==5.4.5
pefile==2023.2.7
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
protobuf==3.6.1
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.16.0
psutil==5.5.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pycairo==1.16.2
pycodestyle==2.11.1
pycrypto==2.6.1
pycryptodome==3.17
pycups==1.9.73
pydeps==1.12.20
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.17.2
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pykerberos==1.1.14
pylint==3.0.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
pysvn==1.9.9
pytest==7.1.2
pytest-cov==3.0.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-slugify==8.0.1
pythondialog==3.4.0
pytimeparse==1.1.5
pytoolconfig==1.2.6
pytz==2023.3.post1
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.5.1
QtPy==2.4.1
reportlab==3.5.34
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resolvelib==0.8.1
rich==13.7.0
roman==2.0.0
rope==1.11.0
rsa==4.7.2
Rtree==1.1.0
s3transfer==0.7.0
seaborn==0.13.0
SecretStorage==2.3.1
selenium==4.21.0
selinux==3.0
simplejson==3.16.0
sip==6.8.0
six==1.14.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==1.8.5
spyder==5.5.0
spyder-kernels==2.5.0
SQLAlchemy==1.3.12
stack-data==0.6.3
stdlib-list==0.10.0
systemd-python==234
testpath==0.4.4
text-unidecode==1.3
textdistance==4.6.0
texttable==1.6.2
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
traitlets==5.14.0
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.12.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.8.0
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==2.2.1
usb-creator==0.3.7
vboxapi==1.0
volatility3==2.4.1
wadllib==1.3.3
watchdog==3.0.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.7.0
whatthepatch==1.0.5
wrapt==1.11.2
wsproto==1.2.0
wurlitzer==3.0.3
xkit==0.0.0
xlrd==1.1.0
xlsx2csv==0.8.2
XlsxWriter==3.2.0
xmltodict==0.12.0
yapf==0.40.2
yara-python==4.3.1
zipp==3.17.0
